python-anpr-service/requirements.txt

flask==2.3.2
flask-cors==4.0.0
opencv-python==4.8.0.74
pytesseract==0.3.10
pillow==10.0.0
numpy==1.24.3
requests==2.31.0
